@@ -1,9 +1,9 @@
 {
-  "name": "semnome017",
+  "name": "zro17",
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "server": "cd server && json-server --watch .\\db.json",
+    "server": "cd server && json-server --watch .\\db.json --routes .\\routes.json --delay 500",
     "build": "cross-env NODE_ENV=production webpack --colors --config ./webpack.prod.js",
     "start": "cross-env NODE_ENV=development webpack-dev-server --colors --open --config ./webpack.dev.js"
   },
@@ -23,26 +23,28 @@
     ]
   },
   "dependencies": {
+    "@ant-design/icons-angular": "^10.2.0",
+    "@antv/g2plot": "^2.3.8",
     "@reach/router": "^1.3.4",
-    "antd": "^4.5.1",
-    "axios": "^0.20.0",
-    "formik": "^2.1.5",
+    "antd": "^4.8.4",
+    "axios": "^0.21.0",
+    "formik": "^2.2.5",
     "formik-antd": "^2.0.1",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-icons": "^3.10.0",
-    "recoil": "0.0.11",
-    "yup": "^0.29.3"
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-waypoint": "^9.0.3",
+    "recoil": "0.1.2",
+    "yup": "^0.30.0"
   },
   "devDependencies": {
     "@types/node": "^12.12.50",
     "@types/reach__router": "^1.3.5",
-    "@types/react": "^16.9.43",
-    "@types/react-dom": "^16.9.8",
+    "@types/react": "^17.0.0",
+    "@types/react-dom": "^17.0.0",
     "@types/react-stickynode": "^3.0.0",
     "@types/recoil": "0.0.1",
     "@types/webpack-merge": "^4.1.5",
-    "@types/yup": "^0.29.4",
+    "@types/yup": "^0.29.9",
     "@typescript-eslint/eslint-plugin": "4.0.1",
     "@typescript-eslint/parser": "4.0.1",
     "clean-webpack-plugin": "^3.0.0",
