@@ -1,16 +1,20 @@
 import React, { FunctionComponent, useState } from 'react';
 import { Link } from '@reach/router';
-import { IoIosClose } from 'react-icons/io';
-import { Button, Drawer } from 'antd';
-import Menu from '../Menu/Menu';
-import MenuAuthentication from '../MenuAuthentication/MenuAuthentication';
+import { Button } from 'antd';
+import MenuPages from '../MenuPages/MenuPages';
+import MenuAuthenticated from '../MenuAuthenticated/MenuAuthenticated';
+import { CloseOutlined } from '@ant-design/icons';
+import { useRecoilValue } from 'recoil';
+import { UserState } from '../../../store/LoginPageState';
+import MenuUnauthenticated from '../MenuUnauthenticated/MenuUnauthenticated';
 
 import './HeaderMobile.scss';
-
-const logo = require('../../../assets/images/logo-alt.svg') as string;
+import Drawer from '../Drawer/Drawer';
 
 const HeaderMobile: FunctionComponent = () => {
   const [sidebarMenuOpen, setSidebarMenuOpen] = useState<boolean>(false);
+  const user = useRecoilValue(UserState);
+  const isAuthenticated = user.Id !== undefined;
 
   const changeDrawer = () => {
     setSidebarMenuOpen(!sidebarMenuOpen);
@@ -20,34 +24,33 @@ const HeaderMobile: FunctionComponent = () => {
     <div className="header-mobile-wrapper default">
       <div className="logo-wrapper">
         <Link to="/">
-          <img src={logo} alt="semnome017" />
-          <h3>semnome017</h3>
+          <h3>zro17</h3>
         </Link>
       </div>
 
-      <Button onClick={changeDrawer} className={`hamburg-btn ${sidebarMenuOpen ? 'active' : ''}`}>
+      <Button onClick={changeDrawer} className={`hamburger ${sidebarMenuOpen ? 'active' : ''}`}>
         <span />
         <span />
         <span />
       </Button>
 
       <Drawer
+        width={285}
         placement="right"
-        closable={false}
-        onClose={changeDrawer}
-        width="285px"
-        className="header-mobile-drawer-wrapper"
-        visible={sidebarMenuOpen}
+        open={sidebarMenuOpen}
+        onClose={() => {
+          setSidebarMenuOpen(false);
+        }}
       >
-        <div className="header-mobile-drawer-close">
-          <button onClick={changeDrawer} type="button">
-            <IoIosClose />
-          </button>
-        </div>
-
-        <div className="main-menu">
-          <MenuAuthentication onNavigate={changeDrawer} />
-          <Menu align="vertical" onNavigate={changeDrawer} />
+        <div className="header-mobile-items-wrapper">
+          {
+            isAuthenticated ? (
+              <MenuAuthenticated onNavigate={changeDrawer} />
+            ) : (
+              <MenuUnauthenticated onNavigate={changeDrawer} />
+            )
+          }
+          <MenuPages align="vertical" onNavigate={changeDrawer} />
         </div>
       </Drawer>
     </div>
