@@ -1,54 +1,54 @@
 import React, { lazy, Suspense } from 'react';
 import ReactDOM from 'react-dom';
 import { RecoilRoot, MutableSnapshot } from 'recoil';
-import { Layout, Spin } from 'antd';
+import { Layout } from 'antd';
 import { Router } from '@reach/router';
-import ProtectedRoute from './components/behavior/ProtectedRoute/ProtectedRoute';
+import { UserState } from './store/LoginPageState';
 import HttpErrorHandling from './components/behavior/HttpErrorHandling/HttpErrorHandling';
 import StateManagement from './components/behavior/StateManagement/StateManagement';
 import ErrorBoundary from './components/behavior/ErrorBoundary/ErrorBoundary';
-import Header from './components/ui/Header/Header';
+
 import User from './domain/User';
-import UserState from './store/atoms/domain/UserState';
+import ScrollToTop from './components/behavior/ScrollToTop/ScrollToTop';
+import ProtectedRoute from './components/behavior/ProtectedRoute/ProtectedRoute';
 
 import './assets/styles/App.scss';
+import Loading from './components/ui/Loading/Loading';
 
 const HomePage = lazy(() => import('./containers/Home/HomePage'));
 const CreateNewAccountPage = lazy(() => import('./containers/Authentication/CreateNewAccount/CreateNewAccountPage'));
 const ForgotPasswordPage = lazy(() => import('./containers/Authentication/ForgotPassword/ForgotPasswordPage'));
 const ResetPasswordPage = lazy(() => import('./containers/Authentication/ResetPassword/ResetPasswordPage'));
 const LoginPage = lazy(() => import('./containers/Authentication/Login/LoginPage'));
-const StocksPage = lazy(() => import('./containers/Search/Stocks/StocksPage'));
-const MePage = lazy(() => import('./containers/Me/MePage'));
-
-const initialize = (snapshot: MutableSnapshot): void => {
-  const store = localStorage.getItem('user');
-
-  if (store !== null) {
-    const user = JSON.parse(store) as User;
-
-    snapshot.set(UserState, user);
-  }
-};
+const CompaniesSearchPage = lazy(() => import('./containers/Companies/Search/CompaniesSearchPage'));
+const CompanyPage = lazy(() => import('./containers/Companies/Company/CompanyPage'));
 
 ReactDOM.render(
-  <RecoilRoot initializeState={initialize}>
+  <RecoilRoot initializeState={(snapshot: MutableSnapshot) => {
+    const user = localStorage.getItem('user');
+
+    if (user !== null) {
+      const loggedUser = JSON.parse(user) as User;
+
+      snapshot.set(UserState, loggedUser);
+    }
+  }}>
     <ErrorBoundary>
       <HttpErrorHandling />
       <StateManagement />
-
       <Layout>
-        <Header />
-        <Suspense fallback={<div />}>
-          <Router>
-            <HomePage default />
-            <HomePage path="/" />
-            <LoginPage path="login" />
-            <StocksPage path="stocks" />
-            <CreateNewAccountPage path="create-new-account" />
-            <ForgotPasswordPage path="forgot-password" />
-            <ResetPasswordPage path="reset-password" />
-            <ProtectedRoute path="me" for={MePage} />
+        <Suspense fallback={<Loading />}>
+          <Router primary={false}>
+            <ScrollToTop path="/">
+              <HomePage default />
+              <HomePage path="/" />
+              <LoginPage path="login" />
+              <CreateNewAccountPage path="create-new-account" />
+              <ForgotPasswordPage path="forgot-password" />
+              <ResetPasswordPage path="reset-password" />
+              <CompaniesSearchPage path="companies" />
+              <ProtectedRoute path="companies/:companyId/:companyName" for={CompanyPage} />
+            </ScrollToTop>
           </Router>
         </Suspense>
       </Layout>
