@@ -5,22 +5,12 @@ const HttpErrorHandling: FunctionComponent = () => {
   axios.interceptors.response.use(undefined, (error: AxiosError) => {
     const statusCode = error.response ? error.response.status : null;
 
-    let message = '';
-
     switch (statusCode) {
-      case 401:
-        message = 'HANDLING - 401';
-        break;
       case 404:
-        message = 'HANDLING - 404';
-        break;
+      case 500:
       default:
-        message = 'HANDLING - DEFAULT';
+        return Promise.reject(error);
     }
-
-    console.error(message);
-
-    return Promise.reject(error);
   });
 
   return null;
