@@ -1,16 +1,23 @@
 import React, { FunctionComponent } from 'react';
+import { useRecoilValue } from 'recoil';
 import { Redirect, RouteComponentProps } from '@reach/router';
+import { UserState } from '../../../store/LoginPageState';
 
 type Props = RouteComponentProps & {
   for: FunctionComponent;
 };
 
-const ProtectedRoute: FunctionComponent<Props> = ({ for: Component }) => {
-  if (false) {
+const ProtectedRoute: FunctionComponent<Props> = ({ 
+  for: Component,
+  location
+}) => {
+  const userState = useRecoilValue(UserState);
+  
+  if (userState.Id !== undefined) {
     return <Component />;
   }
 
-  return <Redirect from="" to="/login" noThrow />;
+  return <Redirect to={`/login?url=${location?.href}`} />;
 };
 
 export default ProtectedRoute;
