@@ -1,5 +1,4 @@
-import React, { Component, FunctionComponent } from 'react';
-
+import React, { Component, CSSProperties, FunctionComponent } from 'react';
 import { Button as ButtonAntd } from 'antd';
 
 import './Button.scss';
@@ -10,7 +9,9 @@ export interface ButtonProps {
   type: 'submit' | 'button' | 'reset',
   design: 'primary' | 'default' | 'link' | 'ghost' | 'dashed' | 'text',
   icon?: Component
+  loading?: boolean,
   disabled?: boolean,
+  style?: CSSProperties,
   onClick?(): void
 }
 
@@ -20,17 +21,20 @@ const Button: FunctionComponent<ButtonProps> = ({
   type,
   design,
   icon,
+  loading,
   disabled,
+  style,
   onClick
 }) => {
   return (
     <ButtonAntd
       className={`btn btn-${size} btn-${design}`}
       type={design}
+      style={style}
       htmlType={type}
       icon={icon}
-      loading={false}
-      disabled={disabled}
+      loading={loading}
+      disabled={loading ? true : disabled}
       onClick={() => {
         if (onClick) {
           onClick();
