@@ -7,8 +7,8 @@ interface WindowSize {
 
 function useWindowSize(): WindowSize {
   const [windowSize, setWindowSize] = useState<WindowSize>({
-    width: 0,
-    height: 0,
+    width: window?.innerWidth ?? 0,
+    height: window?.innerHeight ?? 0,
   });
 
   useEffect(() => {
