@@ -1,7 +1,15 @@
 import Filter from "./Filter";
 
-class FilterGroup {
-  public constructor(params: FilterGroup = {} as FilterGroup) {
+export default class FilterGroup {
+  public constructor(params: Partial<FilterGroup>) {
+    if (!params.Name) {
+      throw new Error('Name must be informed');
+    }
+
+    if (!params.Filters || params.Filters.length < 1) {
+      throw new Error('Filters must be informed');
+    }
+
     this.Name = params.Name;
     this.Filters = params.Filters;
   }
@@ -9,5 +17,3 @@ class FilterGroup {
   Name: string;
   Filters: Array<Filter>;
 }
-
-export default FilterGroup;
