@@ -1,31 +1,38 @@
 import React, { FunctionComponent } from 'react';
 import { Link } from '@reach/router';
-import Menu from '../Menu/Menu';
-import MenuAuthentication from '../MenuAuthentication/MenuAuthentication';
+import MenuPages from '../MenuPages/MenuPages';
+import MenuAuthenticated from '../MenuAuthenticated/MenuAuthenticated';
 
 import './HeaderFullscreen.scss';
+import { useRecoilValue } from 'recoil';
+import { UserState } from '../../../store/LoginPageState';
+import MenuUnauthenticated from '../MenuUnauthenticated/MenuUnauthenticated';
 
-const logo = require('../../../assets/images/logo-alt.svg') as string;
+const HeaderFullscreen: FunctionComponent = () => {
+  const user = useRecoilValue(UserState);
+  const isAuthenticated = user.Id !== undefined;
 
-const HeaderFullscreen: FunctionComponent = () => (
-  <div className="header-fullscreen-wrapper">
-    <div className="navbar-wrapper">
-      <div className="logo-wrapper">
-        <Link to="/">
-          <img src={logo} alt="semnome017" />
-          <h3>semnome017</h3>
-        </Link>
+  return (
+    <div className="header-fullscreen">
+      <Link className="logo-wrapper" to="/">
+        <h3>zro17</h3>
+      </Link>
+
+      <div className="links-wrapper">
+        <MenuPages align="horizontal" />
       </div>
-      <div className="menu-wrapper">
-        <div className="links-wrapper">
-          <Menu align="horizontal" />
-        </div>
-        <div className="auth-wrapper">
-          <MenuAuthentication />
-        </div>
+
+      <div className="auth-wrapper">
+        {
+          isAuthenticated ? (
+            <MenuAuthenticated />
+          ) : (
+            <MenuUnauthenticated />
+          )
+        }
       </div>
     </div>
-  </div>
-);
+  )
+};
 
 export default HeaderFullscreen;
