@@ -1,4 +1,4 @@
-import React, { FunctionComponent } from 'react';
+import React, { FunctionComponent, useState } from 'react';
 import axios from 'axios';
 import Divider from 'antd/lib/divider';
 import { navigate, RouteComponentProps, Link } from '@reach/router';
@@ -10,15 +10,14 @@ import Input from '../../../components/form/Input/Input';
 import Password from '../../../components/form/Password/Password';
 import Switch from '../../../components/form/Switch/Switch';
 import Button from '../../../components/form/Button/Button';
-import UserState from '../../../store/atoms/domain/UserState';
 import User from '../../../domain/User';
+import { UserState } from '../../../store/LoginPageState';
 
 import '../../../assets/styles/Authentication.scss';
 
-const logo = require('../../../assets/images/logo-alt.svg') as string;
-
 const LoginPage: FunctionComponent<RouteComponentProps> = (props: RouteComponentProps) => {
   const [userState, setUserState] = useRecoilState(UserState);
+  const [loading, setLoading] = useState<boolean>(false);
 
   const schema = Yup.object().shape({
     Email: Yup.string()
@@ -31,79 +30,103 @@ const LoginPage: FunctionComponent<RouteComponentProps> = (props: RouteComponent
   });
 
   const submit = async (user: User): Promise<void> => {
-    const response = await axios.get<User>(`${process.env.URL_API_STOCK}login`);
+    setLoading(true);
+
+    const response = await axios.get<User>(`${process.env.URL_API_USER}/login`);
 
     response.data.RememberMe = user.RememberMe;
 
     setUserState(response.data);
 
-    navigate('/');
+    setLoading(false);
+
+    navigate(url());
   };
 
+  const url = (): string => {
+    let route = '/';
+
+    const url = window.location.href.split('?');
+
+    if (url.length > 1) {
+      const params = url[1].split('&');
+
+      for (let param of params) {
+        const pair = param.split('=');
+  
+        const name = pair[0].toLowerCase();
+        const value = pair[1].toLowerCase();
+  
+        if (name === 'url') {
+          route = value;
+          
+          break;
+        }
+      }
+    }
+
+    return route;
+  }
+
   return (
-    <div className="authentication-wrapper">
-      <div className="form-wrapper">
-        <div className="logo-wrapper">
-          <Link to="/">
-            <img src={logo} alt="semnome017" />
-            <h3>semnome017</h3>
-          </Link>
+    <Row className="authentication-wrapper center" gutter={24}>
+      <Col xs={2} sm={2} md={2} lg={6} xl={6}></Col>
+      <Col xs={20} sm={20} md={20} lg={12} xl={12}>
+        <div className="form-wrapper">
+          <div className="title-wrapper">
+            Welcome back to <Link to="/">zro17</Link>
+          </div>
+
+          <div className="subtitle-wrapper">Please log into your account</div>
+
+          <Formik
+            initialValues={{
+              ...userState, Password: ''
+            }}
+            validationSchema={schema}
+            onSubmit={submit}
+          >
+            {(): JSX.Element => (
+              <Form>
+                <Row>
+                  <Col span={24} className="align-left">
+                    <Input name="Email" label="Email" autoComplete="username" value={userState.Email} />
+                  </Col>
+                </Row>
+
+                <Row>
+                  <Col span={24} className="align-left">
+                    <Password name="Password" label="Password" autoComplete="current-password" />
+                  </Col>
+                </Row>
+
+                <Row>
+                  <Col span={12} className="align-left">
+                    <Switch name="RememberMe" label="Remember Me" value={userState.RememberMe} />
+                  </Col>
+                  <Col span={12} className="align-right">
+                    <Link to="/forgot-password">Forgot password?</Link>
+                  </Col>
+                </Row>
+
+                <Row>
+                  <Col span={24} className="align-left">
+                    <Button text="Log In" type="submit" design="primary" size="big" loading={loading} />
+                  </Col>
+                </Row>
+              </Form>
+            )}
+          </Formik>
+
+          <Divider plain>or</Divider>
+          
+          <div className="align-center">
+            <Link to="/create-new-account">Create a new account</Link>
+          </div>
         </div>
-
-        <div className="title-wrapper">Welcome Back</div>
-
-        <div className="title-info-wrapper">Please log into your account</div>
-
-        <Formik
-          initialValues={{
-            ...userState, Password: ''
-          }}
-          validationSchema={schema}
-          onSubmit={submit}
-        >
-          {(): JSX.Element => (
-            <Form>
-              <Row>
-                <Col span={24} className="align-left">
-                  <Input name="Email" label="Email" autoComplete="username" value={userState.Email} />
-                </Col>
-              </Row>
-
-              <Row>
-                <Col span={24} className="align-left">
-                  <Password name="Password" label="Password" autoComplete="current-password" />
-                </Col>
-              </Row>
-
-              <Row>
-                <Col span={12} className="align-left">
-                  <Switch name="RememberMe" label="Remember Me" value={userState.RememberMe} />
-                </Col>
-                <Col span={12} className="align-right">
-                  <Link to="/forgot-password">Forgot password?</Link>
-                </Col>
-              </Row>
-
-              <Row>
-                <Col span={24} className="align-left">
-                  <Button text="Log In" type="submit" design="primary" size="big" />
-                </Col>
-              </Row>
-            </Form>
-          )}
-        </Formik>
-
-        <Divider />
-
-        <div className="align-center">
-          <Link to="/create-new-account">Create a new account</Link>
-        </div>
-      </div>
-
-      <div className="banner-wrapper">
-        <div className="banner-wrapper-image" />
-      </div>
-    </div>
+      </Col>
+      <Col xs={2} sm={2} md={2} lg={6} xl={6}></Col>
+    </Row>
   );
 };
 
