@@ -1,65 +1,92 @@
 import React, { FunctionComponent } from 'react';
 import { RouteComponentProps } from '@reach/router';
+import Header from '../../components/ui/Header/Header';
+import { Content } from '../../components/ui/Content/Content';
 
 const HomePage: FunctionComponent<RouteComponentProps> = (props: RouteComponentProps) => (
-  <div>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-    <p>Home Page</p>
-  </div>
+  <>
+    <Header fixed />
+
+    <Content>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+      <p>home page</p>
+    </Content>
+  </>
 );
 
 export default HomePage;
