@@ -1,11 +1,14 @@
-class User {
-  public constructor(params: User = {} as User) {
+import { Plans } from "./Plans";
+
+export default class User {
+  public constructor(params: Partial<User>) {
     this.Id = params.Id;
-    this.Name = params.Name;
-    this.Email = params.Email;
-    this.RememberMe = params.RememberMe;
-    this.Bearer = params.Bearer;
-    this.Avatar = params.Avatar;
+    this.Name = params.Name ?? "";
+    this.Email = params.Email ?? "";
+    this.RememberMe = params.RememberMe ?? false;
+    this.Bearer = params.Bearer ?? "";
+    this.Avatar = params.Avatar ?? "";
+    this.Plans = params.Plans ?? [];
   }
 
   Id: number | undefined;
@@ -14,6 +17,5 @@ class User {
   RememberMe: boolean;
   Bearer: string;
   Avatar: string;
+  Plans: Plans[];
 }
-
-export default User;
